# apps/api/openapi/avenzo.v1.yaml
openapi: "3.1.0"
info:
  title: Avenzo API
  description: Existing Avenzo backend API contracts. apps/api is the source of truth.
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.avenzo.in
    description: Production
...
paths:
...
  /api/v1/customer/auth/signup:
    post:
      tags: [Auth]
      summary: Customer signup
      operationId: customerSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSignupRequest"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseEnvelope"
        "400":
          $ref: "#/components/responses/V1Error"
        "409":
          $ref: "#/components/responses/V1Error"
  /api/v1/customer/auth/login:
    post:
      tags: [Auth]
      summary: Customer email login
      operationId: customerLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseEnvelope"
        "400":
          $ref: "#/components/responses/V1Error"
        "401":
          $ref: "#/components/responses/V1Error"
        "403":
          $ref: "#/components/responses/V1Error"
  /api/v1/customer/auth/me:
    get:
      tags: [Auth]
      summary: Current customer profile
      operationId: getCustomerAuthMe
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/UserEnvelope"
        "401":
          $ref: "#/components/responses/V1Error"
  /api/v1/restaurant/auth/login:
    post:
      tags: [Auth]
      summary: Restaurant partner login
      operationId: restaurantLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantAuthResponseEnvelope"
        "400":
          $ref: "#/components/responses/V1Error"
        "401":
          $ref: "#/components/responses/V1Error"
        "403":
          $ref: "#/components/responses/V1Error"
  /api/v1/restaurant/me:
    get:
      tags: [Auth]
      summary: Current restaurant user profile
      operationId: getRestaurantMe
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/RestaurantMeEnvelope"
        "401":
          $ref: "#/components/responses/V1Error"
        "403":
          $ref: "#/components/responses/V1Error"
  /api/v1/customer/profile:
    patch:
      tags: [Auth]
      summary: Update customer profile
      operationId: updateCustomerProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerProfileUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/UserEnvelope"
        "400":
          $ref: "#/components/responses/V1Error"
        "401":
          $ref: "#/components/responses/V1Error"
        "403":
          $ref: "#/components/responses/V1Error"
  /api/v1/me:
    get:
      tags: [Auth]
      summary: Generic current user profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/GenericUserEnvelope"
        "401":
          $ref: "#/components/responses/V1Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    V1Error:
      description: Standard v1 error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/V1ErrorEnvelope"
    UserEnvelope:
      description: User profile envelope.
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                const: true
              data:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
    GenericUserEnvelope:
      description: Generic user profile envelope.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/V1Envelope"
    RestaurantMeEnvelope:
      description: Restaurant user and restaurant envelope.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/V1Envelope"
...
  schemas:
    V1Envelope:
      type: object
      properties:
        success:
          const: true
        data:
          type:
            - object
            - array
            - "null"
...
    V1ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type:
            - string
            - "null"
        phone:
          type:
            - string
            - "null"
        role:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CustomerSignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        phone:
          type: string
    CustomerProfileUpdateRequest:
      type: object
      description: At least one field must be provided.
      properties:
        name:
          type:
            - string
            - "null"
          maxLength: 120
        phone:
          type:
            - string
            - "null"
    AuthResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/V1Envelope"
    RestaurantAuthResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/V1Envelope"
...